frontend: Timeago Component
