+++ frontend/packages/core/package.json
@@ -52,6 +52,7 @@
     "react-router": "^6.0.0-beta",
     "react-router-dom": "^6.0.0-beta",
     "react-test-renderer": "^17.0.2",
+    "react-timeago": "^7.0.0",
     "recharts": "^2.1.9",
     "superstruct": "~0.15.0",
     "uuid": "^8.3.2",
