+++ frontend/packages/core/src/index.tsx
@@ -49,6 +49,7 @@ export {
   TreeTable,
 } from "./Table";
 export { default as Code } from "./text";
+export * from "./time";
 export { Typography } from "./typography";
 export { default as ClutchApp } from "./AppProvider";
