+++ frontend/workflows/projectCatalog/package.json
@@ -38,8 +38,7 @@
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-hook-form": "^7.25.3",
-    "react-router-dom": "^6.0.0-beta",
-    "react-timeago": "^7.0.0"
+    "react-router-dom": "^6.0.0-beta"
   },
   "devDependencies": {
     "@clutch-sh/tools": "^2.0.0-beta"
